+++ tests/Support/EasyRule.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Rbac\Tests;
+namespace Yiisoft\Rbac\Tests\Support;
 
 use Yiisoft\Rbac\Item;
 use Yiisoft\Rbac\RuleInterface;
